Phase Two Course is 3d and I tried (not fully succesfully) to shader the different hight levels. therefore we have a working view.
the main in game is right now the only working.
the first time the course starts it will create the render picture, therefore it will take a while. I appology for that. after the second running the course will be fast.
The editor function is right now damaged and a faster prerender algorithm would be needed.

the course is now a Tile[][][] array as my measurments have shown, that fast enough retieval speed should be given and octree would only be interesting, if we dont prerender the playfield.